refactor: :recycle: Refactor NULL check for tokenizer
tokenizer 함수 내부 NULL 체크 추가 및 norm 수정
