+++ makefile
@@ -19,9 +19,10 @@ HGEN     := hgen
 # ===== Packages =====
 PKGS     := prompt lexer parser exec api builtin tree error
 
-lexerV   := lexer expansion tokenizer util util2 \
+lexerV   := lexer expansion util util2 \
 			scanner scanner_list scanner_util \
-			dollar_scan1 dollar_scan2 metachar_scan1 metachar_scan2
+			dollar_scan1 dollar_scan2 metachar_scan1 metachar_scan2 \
+			tokenizer tokenizer_util
 parserV  := parser new1 new2 del util1 util2 expander
 promptV  := prompt interrupt util
 apiV     := shell signal path file
